fix(tests) Fix cyclic dependency introduced to tests
